chore: move layout/list/(page|header) to LayoutListView/components.
